Fix: macro audit issues (#321)
### Description

This PR fixes some of the issues found in the [Macro
audit.](https://www.notion.so/Mento-2-Preliminary-Audit-Report-external-private-9a2a743bddb0450b8e7914a4ccc96704)

We originally planned to fix the following issues but we are only fixing
Q-2, Q-5 and Q-7 in this PR. Reasons are listed below:

[L-1]: Will be fixed by [another
PR](https://github.com/mento-protocol/mento-core/tree/feat/factory-contract-leftovers)
[L-2]: Will be fixed by [another
PR](https://github.com/mento-protocol/mento-core/tree/feat/factory-contract-leftovers)
[Q-2]: ✅ 
[Q-3]: ILocking is not shared between LockingBase and Airgrab as they
stated in the prelim report. Fixing this would require us to introduce a
shared interface and I am not sure if it worths such thing
[Q-4]: Will be fixed by [another
PR](https://github.com/mento-protocol/mento-core/tree/feat/factory-contract-leftovers)
[Q-5]: ✅ 
[Q-7]: ✅ 

As an addition to reviewing fixes, please go through the issues we
decided not to fix and comment below if you disaggree.

### Other changes

Typo that is found by @chapati23 is fixed

### Tested

No breaking changes.

### Related issues

- Fixes #301

+++ contracts/governance/Airgrab.sol
@@ -15,7 +15,7 @@ import { ILocking } from "./locking/interfaces/ILocking.sol";
 /**
  * @title Airgrab
  * @author Mento Labs
- * @notice This contract implements a token airgrab gated by a MerkeTree and KYC using fractal.
+ * @notice This contract implements a token airgrab gated by a MerkleTree and KYC using fractal.
  * The airgrab also forces claimers to immediately lock their tokens as veTokens for a
  * predetermined period.
  * @dev The contract is only Ownable because of the circular dependency
@@ -113,9 +113,9 @@ contract Airgrab is Ownable, ReentrancyGuard {
   }
 
   /**
-   * @dev Check if the account can cliam
+   * @dev Check if the account can claim
    * @notice This modifier checks if the airgrab is still active,
-   * if the account hasn't already claimed and if it's incldued
+   * if the account hasn't already claimed and if it's included
    * in the MerkleTree.
    * @param account The address of the account to check.
    * @param amount The amount of tokens to be claimed.
@@ -128,7 +128,7 @@ contract Airgrab is Ownable, ReentrancyGuard {
   ) {
     require(block.timestamp <= endTimestamp, "Airgrab: finished");
     require(!claimed[account], "Airgrab: already claimed");
-    bytes32 leaf = keccak256(bytes.concat(keccak256(abi.encode(account, uint256(amount)))));
+    bytes32 leaf = keccak256(bytes.concat(keccak256(abi.encode(account, amount))));
     require(MerkleProof.verify(merkleProof, root, leaf), "Airgrab: not in tree");
     _;
   }
